requirements.txt
streamlit==1.56.0
numpy==2.4.4
langchain==1.2.15
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-google-genai==2.1.12
chromadb==1.5.5
pypdf==5.9.0
python-dotenv==1.2.2
